CRM-6687 " Implemented file exist check before unlink "

+++ application/libraries/Booking_utilities.php
@@ -268,7 +268,10 @@ class Booking_utilities {
                 //Update JOb Card Booking
                 $this->My_CI->booking_model->update_booking($booking_id,  array('booking_jobcard_filename'=>$output_file_pdf));
                 $f = TMP_FOLDER.$output_file_pdf;
-                unlink($f);
+                if(file_exists($f))
+                {
+                    unlink($f);
+                }
                 if (!empty($qr) && file_exists($qr)) {
                     $res1 = 0;
                     system(" chmod 777 " . $qr, $res1);
